+++ packages/jaspr_cli/tool/new_component_bricks/new_stateful_component/__brick__/{{name.snakeCase()}}.dart
@@ -1,0 +1,36 @@
+import 'package:jaspr/dom.dart';
+import 'package:jaspr/jaspr.dart';
+{{#client}}
+@client
+{{/client}}
+class {{name.pascalCase()}} extends StatefulComponent {
+  const {{name.pascalCase()}}({super.key});
+
+  @override
+  State<{{name.pascalCase()}}> createState() => _{{name.pascalCase()}}State();
+}
+
+class _{{name.pascalCase()}}State extends State<{{name.pascalCase()}}> {
+
+  {{#client}}
+  @override
+  void initState() {
+    super.initState();
+    if (kIsWeb) {
+      // Run client-side initialization logic here
+    } else {
+      // Run server-side initialization logic here
+    }
+  }
+  {{/client}}
+
+  @override
+  Component build(BuildContext context) {
+    return div([]);
+  }
+
+  {{#styles}}
+  @css
+  List<StyleRule> get styles => [];
+  {{/styles}}
+}